Move RecordableCommands to DynamoCore
