+++ src/DynamoCore/DynamoCore.csproj
@@ -111,6 +111,8 @@ limitations under the License.
     <Compile Include="Core\DebugSettings.cs" />
     <Compile Include="Core\DynamoRunner.cs" />
     <Compile Include="Core\Context.cs" />
+    <Compile Include="Library\RecordableCommands.cs" />
+    <Compile Include="Models\DynamoModelCommands.cs" />
     <Compile Include="Utilities\CodeBlockUtils.cs" />
     <Compile Include="Utilities\NotificationObject.cs" />
     <Compile Include="Core\Threading\AsyncTask.cs" />
