+++ src/DynamoCoreWpf/TestInfrastructure/DirectoryPathMutator.cs
@@ -38,8 +38,8 @@ namespace Dynamo.TestInfrastructure
 
             DynamoViewModel.UIDispatcher.Invoke(new Action(() =>
             {
-                DynamoViewModel.UpdateModelValueCommand updateValue =
-                    new DynamoViewModel.UpdateModelValueCommand(node.GUID, "Value", Environment.CurrentDirectory);
+                UpdateModelValueCommand updateValue =
+                    new UpdateModelValueCommand(node.GUID, "Value", Environment.CurrentDirectory);
 
                 DynamoViewModel.ExecuteCommand(updateValue);
             }));
