+++ src/DynamoCoreWpf/TestInfrastructure/IntegerSliderMutator.cs
@@ -53,8 +53,8 @@ namespace Dynamo.TestInfrastructure
 
                 DynamoViewModel.UIDispatcher.Invoke(new Action(() =>
                 {
-                    DynamoViewModel.UpdateModelValueCommand updateValue =
-                        new DynamoViewModel.UpdateModelValueCommand(node.GUID, "Value", value);
+                    UpdateModelValueCommand updateValue =
+                        new UpdateModelValueCommand(node.GUID, "Value", value);
                     DynamoViewModel.ExecuteCommand(updateValue);
                 }));
