+++ src/DynamoCoreWpf/TestInfrastructure/ListMutator.cs
@@ -41,10 +41,10 @@ namespace Dynamo.TestInfrastructure
                 DynamoViewModel.UIDispatcher.Invoke(new Action(() =>
                 {
                     //create commands
-                    DynamoViewModel.MakeConnectionCommand connToStart1 =
-                        new DynamoViewModel.MakeConnectionCommand(guidNumber1, 0, (PortType)1, (DynamoViewModel.MakeConnectionCommand.Mode)0);
-                    DynamoViewModel.MakeConnectionCommand connToStart2 =
-                        new DynamoViewModel.MakeConnectionCommand(n.GUID, 0, (PortType)0, (DynamoViewModel.MakeConnectionCommand.Mode)1);
+                    MakeConnectionCommand connToStart1 =
+                        new MakeConnectionCommand(guidNumber1, 0, (PortType)1, (MakeConnectionCommand.Mode)0);
+                    MakeConnectionCommand connToStart2 =
+                        new MakeConnectionCommand(n.GUID, 0, (PortType)0, (MakeConnectionCommand.Mode)1);
 
                     //execute commands
                     DynamoViewModel.ExecuteCommand(connToStart1); //"Number" with "List" on InPort(0)
