+++ src/DynamoCoreWpf/ViewModels/NodeViewModel.cs
@@ -565,7 +565,7 @@ namespace Dynamo.ViewModels
 
         private void DeleteNodeAndItsConnectors(object parameter)
         {
-            var command = new DynamoViewModel.DeleteModelCommand(nodeLogic.GUID);
+            var command = new DeleteModelCommand(nodeLogic.GUID);
             DynamoViewModel.ExecuteCommand(command);
         }
